2020-12-27 back up

+++ oh-my-zsh/plugins/puddletown-apt/puddletown-apt.plugin.zsh
@@ -1,5 +1,8 @@
 # Linux Apt package manager shortcuts
-alias ai="sudo apt update && sudo apt install -y"
-alias au="sudo apt purge"
-alias as="sudo apt update && apt search"
-alias update-linux="sudo apt update && sudo apt upgrade && sudo apt full-upgrade -y"
+if [[ $OSTYPE==*linux* ]]; then
+	alias ai="sudo apt update && sudo apt install -y"
+	alias au="sudo apt purge"
+	alias as="sudo apt update && apt search"
+	alias update-linux="sudo apt update && sudo apt upgrade && sudo apt full-upgrade -y"
+fi
+
